update resnet50 quantization e2e example
