+++ onnxruntime/python/tools/quantization/onnx_model.py
@@ -50,6 +50,10 @@ class ONNXModel:
     def remove_initializer(self, tensor):
         if tensor in self.model.graph.initializer:
             self.model.graph.initializer.remove(tensor)
+            for input in self.model.graph.input:
+                if input.name == tensor.name:
+                    self.model.graph.input.remove(input)
+                    break
 
     def remove_initializers(self, init_to_remove):
         for initializer in init_to_remove:
@@ -126,7 +130,68 @@ class ONNXModel:
                 if node_input == initializer.name:
                     nodes.append(node)
         return nodes
-    
+       
+    def replace_gemm_with_matmul(self):
+        new_nodes = []
+
+        for node in self.nodes():
+            if node.op_type == 'Gemm':
+                alpha = 1.0
+                beta = 1.0
+                transA = 0
+                transB = 0
+                for attr in node.attribute:
+                    if attr.name == 'alpha':
+                        alpha = onnx.helper.get_attribute_value(attr)
+                    elif attr.name == 'beta':
+                        beta = onnx.helper.get_attribute_value(attr)
+                    elif attr.name == 'transA':
+                        transA = onnx.helper.get_attribute_value(attr)
+                    elif attr.name == 'transB':
+                        transB = onnx.helper.get_attribute_value(attr)
+                if alpha == 1.0 and beta == 1.0 and transA == 0:
+                    inputB = node.input[1]
+                    if transB == 1:
+                        B = self.get_initializer(node.input[1])
+                        if B:
+                            # assume B is not used by any other node
+                            B_array = onnx.numpy_helper.to_array(B)
+                            B_trans = onnx.numpy_helper.from_array(B_array.T)
+                            B_trans.name = B.name
+                            self.remove_initializer(B)
+                            self.add_initializer(B_trans)
+                        else:
+                            inputB += '_Transposed'
+                            transpose_node = onnx.helper.make_node('Transpose',
+                                                                inputs=[node.input[1]],
+                                                                outputs=[inputB],
+                                                                name=node.name+'_Transpose')
+                            new_nodes.append(transpose_node)
+
+                    matmul_node = onnx.helper.make_node('MatMul',
+                                                        inputs=[node.input[0], inputB],
+                                                        outputs=[node.output[0] + ('_MatMul' if len(node.input)>2 else '')],
+                                                        name=node.name + '_MatMul')
+                    new_nodes.append(matmul_node)
+
+                    if len(node.input) > 2:
+                        add_node = onnx.helper.make_node('Add',
+                                                         inputs=[node.output[0] + '_MatMul', node.input[2]],
+                                                         outputs=node.output,
+                                                         name=node.name + '_Add')
+                        new_nodes.append(add_node)  
+                
+                # unsupported
+                else:
+                    new_nodes.append(node)
+            
+            # not GEMM
+            else:
+                new_nodes.append(node)
+
+        self.graph().ClearField('node')
+        self.graph().node.extend(new_nodes)
+
     def save_model_to_file(self, output_path, use_external_data_format=False):
         '''
         Save model to external data, which is needed for model size > 2GB
